requirements.txt
#devops
black
click
pytest
pytest-cov
pylint
#rust based linter
ruff
boto3
#web
fastapi
uvicorn
#datascience
jupyter
pandas
numpy
polars
scikit-learn
matplotlib
seaborn
dask
python-dotenv
databricks-sql-connector
#openai
openai
#flask
Flask